allow setting reindexing time. fixes #7
This removes cron to trigger reindxing which has two problems:

1) it runs every 10 minutes regardless of how long the indexing takes
2) it requires a rather complicated syntax which make making the time
   configurable somewhat complicated

This replaces the mechanism with a simple sleep within a while loop.

The reindexing period can be set by passing a REINDEX environment
varible. It expects an integer defining the number of minutes between
indexing. Setting it to 0 will disabling automatic reindexing.